+++ utils/types.ts
@@ -12,7 +12,26 @@ export interface SharedModalProps {
 
 
 export interface AlbumItem {
+  uploader: string;
   image: string;
   video?: boolean;
+  duration?: number; // In seconds
   contentTimeStamp: number;
+  likes?: string[];
+  hidden?: boolean;
+  private?: boolean;
+  deleted?: boolean;
+  hiddenFor?: string[];
+  batch?: string;
+  firstOf?: number;
+  cachedCommentsCount?: number;
+  contentLocationHash?: string; // precision of 9!
+  timeStamp: number;
+  lastChange: number;
+  viewCount: { [userId: string]: number };
+  unreadItems?: { [userId: string]: number };
+
+  sha1KB?: string;
+  imageHash?: string;
+  rating?: number;
 }